refactor: changed test structure
