+++ meson.build
@@ -99,7 +99,7 @@ gtest_dep = dependency('gtest', version: '>=1.10.0', fallback: ['gtest', 'gtest_
 gmock_main_dep = dependency('gmock_main', version: '>=1.10.0', fallback: ['gtest', 'gmock_main_dep'])
 
 fscheck = executable('fs_check',
-	'test/webfuse/tests/integration/fs_check.c')
+	'test/webfuse/fs_check/main.c')
 
 openssl = find_program('openssl')
 test_server_certs = custom_target('test_server_certs',
@@ -113,17 +113,17 @@ test_certs_dep = declare_dependency(
 	sources: [test_server_certs, test_client_certs])
 
 alltests = executable('alltests',
-	'test/webfuse/tests/core/jsonrpc/mock_timer_callback.cc',
-	'test/webfuse/tests/core/jsonrpc/mock_timer.cc',
-	'test/webfuse/tests/core/jsonrpc/test_is_request.cc',
-	'test/webfuse/tests/core/jsonrpc/test_request.cc',
-	'test/webfuse/tests/core/jsonrpc/test_is_response.cc',
-	'test/webfuse/tests/core/jsonrpc/test_response.cc',
-	'test/webfuse/tests/core/jsonrpc/test_server.cc',
-	'test/webfuse/tests/core/jsonrpc/test_proxy.cc',
-	'test/webfuse/tests/core/jsonrpc/test_response_parser.cc',
-	'test/webfuse/tests/core/timer/test_timepoint.cc',
-	'test/webfuse/tests/core/timer/test_timer.cc',
+	'test/webfuse/jsonrpc/mock_timer_callback.cc',
+	'test/webfuse/jsonrpc/mock_timer.cc',
+	'test/webfuse/jsonrpc/test_is_request.cc',
+	'test/webfuse/jsonrpc/test_request.cc',
+	'test/webfuse/jsonrpc/test_is_response.cc',
+	'test/webfuse/jsonrpc/test_response.cc',
+	'test/webfuse/jsonrpc/test_server.cc',
+	'test/webfuse/jsonrpc/test_proxy.cc',
+	'test/webfuse/jsonrpc/test_response_parser.cc',
+	'test/webfuse/timer/test_timepoint.cc',
+	'test/webfuse/timer/test_timer.cc',
 	'test/webfuse/test_util/tempdir.cc',
 	'test/webfuse/test_util/file_utils.cc',
 	'test/webfuse/test_util/timeout_watcher.cc',
@@ -136,31 +136,31 @@ alltests = executable('alltests',
 	'test/webfuse/mocks/mock_operation_context.cc',
 	'test/webfuse/mocks/mock_jsonrpc_proxy.cc',
 	'test/webfuse/mocks/mock_adapter_client_callback.cc',
-	'test/webfuse//tests/core/test_util.cc',
-	'test/webfuse/tests/core/test_container_of.cc',
-	'test/webfuse/tests/core/test_slist.cc',
-	'test/webfuse/tests/core/test_base64.cc',
-	'test/webfuse/tests/core/test_status.cc',
-	'test/webfuse/tests/core/test_message.cc',
-	'test/webfuse/tests/core/test_message_queue.cc',
-	'test/webfuse/tests/core/test_url.cc',
-	'test/webfuse/tests/adapter/test_server.cc',
-	'test/webfuse/tests/adapter/test_server_config.cc',
-	'test/webfuse/tests/adapter/test_credentials.cc',
-	'test/webfuse/tests/adapter/test_authenticator.cc',
-	'test/webfuse/tests/adapter/test_authenticators.cc',
-	'test/webfuse/tests/adapter/test_mountpoint.cc',
-	'test/webfuse/tests/adapter/test_fuse_req.cc',
-	'test/webfuse/tests/adapter/operation/test_context.cc',
-	'test/webfuse/tests/adapter/operation/test_open.cc',
-	'test/webfuse/tests/adapter/operation/test_close.cc',
-	'test/webfuse/tests/adapter/operation/test_read.cc',
-	'test/webfuse/tests/adapter/operation/test_readdir.cc',
-	'test/webfuse/tests/adapter/operation/test_getattr.cc',
-	'test/webfuse/tests/adapter/operation/test_lookup.cc',
-	'test/webfuse/tests/integration/file.cc',
-	'test/webfuse/tests/adapter/test_client.cc',
-	'test/webfuse/tests/adapter/test_client_tlsconfig.cc',
+	'test/webfuse/util/test_util.cc',
+	'test/webfuse/util/test_container_of.cc',
+	'test/webfuse/util/test_slist.cc',
+	'test/webfuse/util/test_base64.cc',
+	'test/webfuse/util/test_url.cc',
+	'test/webfuse/test_status.cc',
+	'test/webfuse/test_message.cc',
+	'test/webfuse/test_message_queue.cc',
+	'test/webfuse/test_server.cc',
+	'test/webfuse/test_server_config.cc',
+	'test/webfuse/test_credentials.cc',
+	'test/webfuse/test_authenticator.cc',
+	'test/webfuse/test_authenticators.cc',
+	'test/webfuse/test_mountpoint.cc',
+	'test/webfuse/test_fuse_req.cc',
+	'test/webfuse/operation/test_context.cc',
+	'test/webfuse/operation/test_open.cc',
+	'test/webfuse/operation/test_close.cc',
+	'test/webfuse/operation/test_read.cc',
+	'test/webfuse/operation/test_readdir.cc',
+	'test/webfuse/operation/test_getattr.cc',
+	'test/webfuse/operation/test_lookup.cc',
+	'test/webfuse/integration/file.cc',
+	'test/webfuse/test_client.cc',
+	'test/webfuse/test_client_tlsconfig.cc',
 	link_args: [
 		'-Wl,--wrap=wf_impl_timer_manager_create',
 		'-Wl,--wrap=wf_impl_timer_manager_dispose',
